Updated to 1.8.7

+++ PdfBoxAndroid/src/org/apache/pdfbox/pdmodel/common/function/type4/ExecutionContext.java
@@ -57,7 +57,7 @@ public class ExecutionContext
      */
     public int popInt()
     {
-        return ((Integer)stack.pop()).intValue();
+        return ((Integer)stack.pop());
     }
 
     /**
